Move DummyCanvasFactory to test code as it's only used in tests
